+++ vadr_seed.pm
@@ -1,6 +1,6 @@
 #!/usr/bin/env perl
 # 
-# version: 1.6 [Nov 2023]
+# version: 1.6.1 [Nov 2023]
 #
 # vadr_seed.pm
 # Eric Nawrocki